// backend/src/enums/order-list.enum.ts
export enum OrderListEnum {
    ID = "id",
    _ID = "-id",
    NAME = "name",
    _NAME = "-name",
    SURNAME = "surname",
    _SURNAME = "-surname",
    EMAIL = "email",
    _EMAIL = "-email",
    PHONE = "phone",
    _PHONE = "-phone",
    AGE = "age",
    _AGE = "-age",
    COURSE = "course",
    _COURSE = "-course",
    COURSE_FORMAT = "course-format",
    _COURSE_FORMAT = "-course-format",
    COURSE_TYPE = "course-type",
    _COURSE_TYPE = "-course-type",
    STATUS = "status",
    _STATUS = "-status",
    SUM = "sum",
    _SUM = "-sum",
    ALREADY_PAID = "already-paid",
    _ALREADY_PAID = "-already-paid",
    GROUP = "group",
    _GROUP = "-group",
    CREATED_AT = "created-at",
    _CREATED_AT = "-created-at",
    MANAGER = "manager",
    _MANAGER = "-manager",

}